requirements.txt
Django==2.1.7
djangorestframework==3.9.2
psycopg2==2.7.7
pytz==2018.9
psycopg2-binary
whitenoise==3.3.1
gunicorn==19.9.0
django-rest-swagger==2.2.0
cx_Oracle==6.3.1

requests
pyyaml
tldextract
argparse